requirements.txt
Django~=4.1.7
djangorestframework~=3.14.0
numpy~=1.24.2
pandas~=1.5.3
django-cors-headers==3.14.0
django-pandas~= 0.6.6
djangorestframework-simplejwt==5.2.2


python-dateutil~=2.8.2